backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis[hiredis]==5.2.1
httpx==0.28.1
python-json-logger==3.2.1
stripe==11.6.0
email-validator==2.2.0
orjson==3.10.15
sentry-sdk[fastapi]==2.24.1
pytest==8.3.4
pytest-asyncio==0.25.3
ruff==0.9.4
